LayoutTests/fast/dom/htmlcollection-selectedOptions-namedItem-crash.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Makes sure HTMLCollection.namedItem() does not crash for collection of selected options.");
debug("This test passes if it does not crash.");

var select = document.createElement("select");
var options = select.selectedOptions;
select.length = 10;
shouldBeNull('options.namedItem("test")');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
